Fairly substantial edits to the `Rmd`.  Added content to the `yaml` header to help render the Word document better.  Added keywords, text edits throughout.  Noting generally substantial, mostly tinkering to get it so it sounds ready for a journal.
Rendered to `md`.

Added publications to the `bib` file.

Some changes to the `static_page` document, but mostly to clean up the code.

Changed the way that the `RODBC` connection is made in `assign_fun`, Using an explicit string now so that we don't need to much around with the stuff on the Penn State server.

Added the Description node, but it's giving me troubles.  Not sure how to resolve that.

Added comments in the `processing_code`.