`processing_code` adds a new chunk with the driver obscured.  Added the output chunk but still havent written the SQL code needed.  Will write it on my own computer & push/pull.
Still need to implement the date scanning.

`sql_calls` was choking on FAUNMAP records since they didn't have PIs for most records.  Changed the `INNER JOIN` to a `LEFT JOIN` and fixed the problem.

`assign_fun` changes back to use the `doi_sens` file, and modifies the `description` chunks of the XML.  I wasn't reading the documentation properly, and kept getting errors when validating the schema.  Needed to set `documentationType` as an attrbiute, not a node.

Cleaned up `outputs.csv` again.