Big push to CEI2:
Added the date checking module to the head of `processing_code`.  Added libraries up front.

Used some `dplyr` magic in the hopes of speeding things up.

Removed the copy & pasting of files across the file system.  I'm rendering the files in place now.  `unlink` and creation of the symlink also frees up space around the `js` files created by the `rmarkdown` generator.

The `assign_fun` has some new code to read in the `doi_sens` file.  Added the path to the files into `doi_sens`, which may have been a mistake.  We need to make sure that it's there on CEI2.

`assign_fun` does seem to be failing on datasets using UTF-8 (I think).

Changes to the `static_page` are minimal, but pushes things out to `doi_sens`, make the output of `outputs.csv` an explicit data.frame, since it was causing weirdness before.

I've added my local version of `outputs.csv` here, to make sure we're skipping where we ought to.