Cleaned up the header, but still don't seem to get the favicon. . .
Cleared out `outputs.csv`

Changed the shoulder in the `assign_fun.R`.

Changed the base tiles for the `leaflet` map.  Moved the marker to the top left of a polygon.  Increased the zoom.

Fixed the authorship, and modified it if the publication comes from FAUNMAP, so that the `researcher` is recognized.

Small change in the `processing_code` for the dry run, just running the first three records.