Added batch file (`run_doi.bat`) to attach to the scheduled task, but it is ignored for the repository.
`builder/static_page.csv` has been heavily re-written to link to new libraries, have new output status and to link to local files in the virtual directory for Neotoma.  It also now does error checking to see if the DOI has already been assigned.

`builder/R/assign_fun.R` has now been updated to the proper Neotoma shoulder, as opposed to the sandbox shoulder.

Processing code links to the pandoc path in the Neotoma Server directory.  it also finds `doi_sens` in the root working directory.  This is a change to accmodate the new working path.